refactor:モード選択画面再リファクタリング #175
