+++ front/src/app/select-mode/SelectMode.module.scss
